Finished register, login, logout and CRUD operations for posts

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +16,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate());
+        $user = request()->user();
+        $posts = $user->posts()->with('author')->paginate();
+        return PostResource::collection($posts);
     }
 
     /**
@@ -24,7 +27,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id;
 
         $post = Post::create($data);
         return response()->json(new PostResource($post), 201);
@@ -35,6 +38,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $user = request()->user();
+        if($user->id != $post->author_id){
+            abort(403, 'Access Forbidden');
+        }
         return response()->json(new PostResource($post));
     }
 
@@ -43,8 +50,12 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $user = request()->user();
+        if($user->id != $post->author_id){
+            abort(403, 'Access Forbidden');
+        }
         $data = $request->validate([
-            'title' => ['required', 'string', 'min"2'],
+            'title' => ['required', 'string', 'min:2'],
             'body' => ['required', 'string', 'min:2']
         ]);
 
@@ -58,6 +69,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $user = request()->user();
+        if($user->id != $post->author_id){
+            abort(403, 'Access Forbidden');
+        }
         $post->delete();
         return response()->noContent();
     }
